+++ app/Http/Controllers/PermissionController.php
@@ -13,6 +13,9 @@ class PermissionController extends Controller
 {
     public function index(Request $request)
     {
+        if (!auth()->user()->can('ViewPermissions')) {
+            abort(403, 'Unauthorized action');
+        }
         if ($request->ajax()) {
             $permissions = Permission::query();
             return DataTables::of($permissions)
@@ -20,8 +23,15 @@ class PermissionController extends Controller
                     return Carbon::parse($permission->updated_at)->format('Y-m-d H:i:s');
                 })
                 ->addColumn('action', function ($permission) {
-                    $editIcon = '<a href="' . route('permission.edit', $permission->id) . '" class="text-warning" ><i class="fas fa-edit"></i></a>';
-                    $deleteIcon = '<a href="#" class="text-danger delete-btn" data-id="' . $permission->id . '" ><i class="fas fa-trash-alt"></i></a>';
+                    $editIcon = '';
+                    $deleteIcon = '';
+
+                    if (auth()->user()->can('EditPermissions')){
+                        $editIcon = '<a href="' . route('permission.edit', $permission->id) . '" class="text-warning" ><i class="fas fa-edit"></i></a>';
+                    }
+                    if (auth()->user()->can('DeletePermissions')){
+                        $deleteIcon = '<a href="#" class="text-danger delete-btn" data-id="' . $permission->id . '" ><i class="fas fa-trash-alt"></i></a>';
+                    }
 
                     return '<div class="action_icon">' . $editIcon . $deleteIcon . '</div>';
                 })
@@ -37,11 +47,17 @@ class PermissionController extends Controller
     //create employee and store
     public function create()
     {
+        if (!auth()->user()->can('CreatePermissions')) {
+            abort(403, 'Unauthorized action');
+        }
         return view('permission.create');
     }
 
     public function store(StorePermission $request)
     {
+        if (!auth()->user()->can('CreatePermissions')) {
+            abort(403, 'Unauthorized action');
+        }
         $data = $request->validated();
 
         Permission::create($data);
@@ -52,11 +68,18 @@ class PermissionController extends Controller
     //edit and update
     public function edit(Permission $permission)
     {
+        if(!auth()->user()->can('EditPermissions')) {
+            abort(403, 'Unauthorized action');
+        }
         return view('permission.edit', compact('permission'));
     }
 
     public function update(UpdatePermission $request, Permission $permission)
     {
+        if (!auth()->user()->can('EditPermissions')) {
+            abort(403, 'Unauthorized action');
+        }
+
         $data = $request->validated();
         $permission->update($data);
         return redirect()->route('permission.index')->with('updated', "Permission is successfully updated.");
@@ -65,6 +88,9 @@ class PermissionController extends Controller
     //delete
     public function destroy(Permission $permission)
     {
+        if(!auth()->user()->can('DeletePermissions')) {
+            abort(403, 'Unauthorized action');
+        }
         $permission->delete();
         return 'success';
     }
