+++ app/Http/Controllers/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreRole;
 use App\Http\Requests\UpdateRole;
+use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
@@ -15,24 +16,32 @@ class RoleController extends Controller
 {
     public function index(Request $request)
     {
+        if (!auth()->user()->can('ViewRoles')) {
+            abort(403, 'Unauthorized action');
+        }
         if ($request->ajax()) {
             $roles = Role::query();
-//            $permissions = Permission::query();
             return DataTables::of($roles)
                 ->editColumn('updated_at', function ($role) {
                     return Carbon::parse($role->updated_at)->format('Y-m-d H:i:s');
                 })
-                ->addColumn('permission', function ($role){
+                ->addColumn('permission', function ($role) {
                     $output = '';
-                    foreach($role->permissions as $permission){
-                        $output .= '<span class="badge badge-pill badge-primary m-1">'.$permission->name.'</span>';
+                    foreach ($role->permissions as $permission) {
+                        $output .= '<span class="badge badge-pill badge-primary m-1">' . $permission->name . '</span>';
                     }
                     return $output;
                 })
                 ->addColumn('action', function ($role) {
-                    $editIcon = '<a href="' . route('role.edit', $role->id) . '" class="text-warning" ><i class="fas fa-edit"></i></a>';
-                    $deleteIcon = '<a href="#" class="text-danger delete-btn" data-id="' . $role->id . '" ><i class="fas fa-trash-alt"></i></a>';
+                    $editIcon = '';
+                    $deleteIcon = '';
 
+                    if (auth()->user()->can('EditRoles')) {
+                        $editIcon = '<a href="' . route('role.edit', $role->id) . '" class="text-warning" ><i class="fas fa-edit"></i></a>';
+                    }
+                    if (auth()->user()->can('DeleteRoles')) {
+                        $deleteIcon = '<a href="#" class="text-danger delete-btn" data-id="' . $role->id . '" ><i class="fas fa-trash-alt"></i></a>';
+                    }
                     return '<div class="action_icon">' . $editIcon . $deleteIcon . '</div>';
                 })
                 ->addColumn('plusIcon', function ($role) {
@@ -47,15 +56,21 @@ class RoleController extends Controller
     //create employee and store
     public function create()
     {
+        if (!auth()->user()->can('CreateRoles')) {
+            abort(403, 'Unauthorized action');
+        }
         $permissions = Permission::all();
         return view('role.create', compact('permissions'));
     }
 
     public function store(StoreRole $request)
     {
+        if (!auth()->user()->can('CreateRoles')) {
+            abort(403, 'Unauthorized action');
+        }
         $data = $request->validated();
 
-        $role =Role::create($data);
+        $role = Role::create($data);
         $role->givePermissionTo($request->permissions);
         return redirect()->route('role.index')->with('create', 'Role is successfully created.');
     }
@@ -63,6 +78,9 @@ class RoleController extends Controller
     //edit and update
     public function edit(Role $role)
     {
+        if (!auth()->user()->can('EditRoles')) {
+            abort(403, 'Unauthorized action');
+        }
         $permissions = Permission::all();
         $oldPermissions = $role->permissions->pluck('id')->toArray();
         return view('role.edit', compact('role', 'oldPermissions', 'permissions'));
@@ -70,6 +88,9 @@ class RoleController extends Controller
 
     public function update(UpdateRole $request, Role $role)
     {
+        if (!auth()->user()->can('EditRoles')) {
+            abort(403, 'Unauthorized action');
+        }
         $oldPermissions = $role->permissions->pluck('name')->toArray();
         $role->revokePermissionTo($oldPermissions);
 
@@ -82,6 +103,9 @@ class RoleController extends Controller
     //delete
     public function destroy(Role $role)
     {
+        if (!auth()->user()->can('DeleteRoles')) {
+            abort(403, 'Unauthorized action');
+        }
         $role->delete();
         return 'success';
     }
